Switches to CJS in examples
While ES modules are nice, they're currently not supported on latest
Node and they transpile a bit complicated on latest babel. In order to
simplify the example so one can get running as fast as possible we can
just switch over to using plain CommonJS modules which are simple enough
to understand.