Relaxes the spec requirements for afterResponse
Spec currently states that we should call String/toString on the
response after we've called afterResponse method. This change is to
relax those requirements enabling you to pass whatever type you'd like.
The end result should still be a string (of HTML ideally) but I don't
think we have to specify that.