Fixes exception handling stack traces
On some cases we pass/throw a non-error, we used to wrap that in an
Error and then serialize that new error. The problem with that approach
is that the stack trace would be completely lost.

Now we either include the original stack trace or no stack trace at all.
Because an incorrect stack trace is worse than no stack trace.