If invalid HTML is returned we reject with an err
As a safeguard -- in case a developer pops out of the promise chain an
error happens and Hypernova is unable to catch it -- we can detect whether
something is "off" by inspecting `html`. The html returned should be a
string and exist. If it does not exist then that's a sign that something
is wrong and needs investigation. This should prevent errors from
slipping through the cracks.