Allow a logger instance to be injected via the configuration
Currently, config can specify the options to be passed to the
constructed winston logger, but the specific transport is hard coded as
Console.  Some applications may wish to use other transports like Syslog
that are defined in other packages.

This change adds a config field, `loggerInstance`, that allows an
arbitrarily constructed logger to be passed in.  Really any object that
implements `log` should work, but this is intended for other winston
loggers.