Execute jobs in a batch in serial
Previously, we dispatched all jobs in a batch concurrently.  Rendering
jobs are compute bound, so we would not actually get parallel execution.
One effect of this is that timers for a job in plugins record a start
time from the point where all jobs are dispatched and record the end
time after all or most of the jobs have completely finished.  This
prevents us from being able to reliably tell how much latency a single
job contributes to the total latency (for instance, if job A takes
500ms, and job B and C each take 25ms, we may get results that would
indicate that jobs A, B, and C each spent 550ms executing, which is true
in a way, but unhelpful).

This change executes the jobs serially as a chain.

This should have no or minimal impact either way on actual response
time.