Refactor hypernova server into subcomponents
The previous version is highly tied to a cluster-based
coordinator/worker model with some custom lifecycle management.  To
allow experimentation with other process/configuration management, it is
helpful to break up and export some of the components of the server
configuration.  This preserves the old interface, but allows an a la
carte option.