Fix specs: chai API issues => async test issues =>  ensure tests are running
`isStrictEqual` is not part of the API: replaced with `strictEqual`
Ensure promise is either rejected or resolved correctly.

This was entirely missed because I forgot to ensure the tests were running. I was breaking the test suite by removing code from StrictPromise -- but not ensuring the strict-promise-test was failing appropriately.