Add "<" (&lt) to the ENCODE list
We have found that browsers (at least Chrome and Firefox)
can interpret "/" to close a tag in a HTML comment in a
script tag.

This means that a malicious attacker could submit a
payload like "</script/", prematurely closing the script
tag that a Hypernova comment is in, causing the page to
crash.